VC2010-379894818545237f9dacdbe0ebdd41bb3a85efbc

+++ cl_dll/hud_redraw.cpp
@@ -205,7 +205,7 @@ int CHud :: Redraw( float flTime, int intermission )
 	if ( g_iVisibleMouse )
 	{
 		void IN_GetMousePos( int *mx, int *my );
-		float mx, my;
+		int mx, my;
 
 		IN_GetMousePos( &mx, &my );
